+++ CMakeLists.txt
@@ -35,10 +35,11 @@ include(GNUInstallDirs)
 #Subdirectories
 include_directories(${CMAKE_SOURCE_DIR})
 include_directories("services")
+add_subdirectory("Frontend/GFXUtilities")
 add_subdirectory("Frontend/Graphics")
 add_subdirectory("Frontend/GItems")
-add_subdirectory("Frontend/GFXUtilities")
 add_subdirectory("Frontend/GUI")
+add_subdirectory("Frontend/RUGraph")
 add_subdirectory("Frontend/GLayouts")
 
 set(MAIN_SRC_FILES
@@ -50,7 +51,7 @@ add_library(${PROJECT_NAME} SHARED ${MAIN_SRC_FILES})
 
 #Link libraries
 target_link_libraries(${PROJECT_NAME}
-	GFX GPanel GLayout GUI GFXItem GU GNet DB)#custom libs
+	GFX GPanel GLayout GraphGUI GUI GFXItem GU GNet DB)#custom libs
 
 target_include_directories(gfxplusplus PRIVATE "include")
 target_include_directories(gfxplusplus PRIVATE "Frontend")
@@ -60,14 +61,14 @@ target_include_directories(gfxplusplus PRIVATE "Frontend")
 #    $<INSTALL_INTERFACE:include>
 #    PRIVATE Frontend)
 
-install(TARGETS ${PROJECT_NAME} EXPORT gfxConfig
+install(TARGETS ${PROJECT_NAME} EXPORT gfxplusplusConfig
     ARCHIVE  DESTINATION ${CMAKE_INSTALL_LIBDIR}
     LIBRARY  DESTINATION ${CMAKE_INSTALL_LIBDIR}
     RUNTIME  DESTINATION ${CMAKE_INSTALL_BINDIR})
 
-install(EXPORT gfxConfig DESTINATION share/glades/cmake)
+install(EXPORT gfxplusplusConfig DESTINATION share/gfxplusplus/cmake)
 
-export(TARGETS ${PROJECT_NAME} FILE gfxConfig.cmake)
+export(TARGETS ${PROJECT_NAME} FILE gfxplusplusConfig.cmake)
 
 #make fmt
 add_custom_target(fmt ALL
