+++ Frontend/GFXUtilities/CMakeLists.txt
@@ -10,11 +10,11 @@ set(GU_src_files
 )
 add_library(GU ${GU_src_files})
 
-install(TARGETS GU EXPORT gfxConfig
+install(TARGETS GU EXPORT gfxplusplusConfig
     ARCHIVE  DESTINATION ${CMAKE_INSTALL_LIBDIR}
     LIBRARY  DESTINATION ${CMAKE_INSTALL_LIBDIR}
     RUNTIME  DESTINATION ${CMAKE_INSTALL_BINDIR})
 
-install(EXPORT gfxConfig DESTINATION share/glades/cmake)
+install(EXPORT gfxplusplusConfig DESTINATION share/gfxplusplus/cmake)
 
-export(TARGETS GU FILE gfxConfig.cmake)
+export(TARGETS GU FILE gfxplusplusConfig.cmake)
