+++ Frontend/GLayouts/CMakeLists.txt
@@ -6,11 +6,11 @@ set(GLayout_src_files
 )
 add_library(GLayout ${GLayout_src_files})
 
-install(TARGETS GLayout EXPORT gfxConfig
+install(TARGETS GLayout EXPORT gfxplusplusConfig
     ARCHIVE  DESTINATION ${CMAKE_INSTALL_LIBDIR}
     LIBRARY  DESTINATION ${CMAKE_INSTALL_LIBDIR}
     RUNTIME  DESTINATION ${CMAKE_INSTALL_BINDIR})
 
-install(EXPORT gfxConfig DESTINATION share/glades/cmake)
+install(EXPORT gfxplusplusConfig DESTINATION share/gfxplusplus/cmake)
 
-export(TARGETS GLayout FILE gfxConfig.cmake)
+export(TARGETS GLayout FILE gfxplusplusConfig.cmake)
