+++ Frontend/Graphics/CMakeLists.txt
@@ -8,11 +8,11 @@ set(GFX_src_files
 )
 add_library(GFX ${GFX_src_files})
 
-install(TARGETS GFX EXPORT gfxConfig
+install(TARGETS GFX EXPORT gfxplusplusConfig
     ARCHIVE  DESTINATION ${CMAKE_INSTALL_LIBDIR}
     LIBRARY  DESTINATION ${CMAKE_INSTALL_LIBDIR}
     RUNTIME  DESTINATION ${CMAKE_INSTALL_BINDIR})
 
-install(EXPORT gfxConfig DESTINATION share/glades/cmake)
+install(EXPORT gfxplusplusConfig DESTINATION share/gfxplusplus/cmake)
 
-export(TARGETS GFX FILE gfxConfig.cmake)
+export(TARGETS GFX FILE gfxplusplusConfig.cmake)
