+++ Frontend/RUGraph/Circle.cpp
@@ -15,8 +15,8 @@
 // DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include "Circle.h"
-#include "../../GFXUtilities/point2.h"
-#include "../../Graphics/graphics.h"
+#include "../GFXUtilities/point2.h"
+#include "../Graphics/graphics.h"
 #include "RUGraph.h"
 
 Circle::Circle(RUGraph* newParent, SDL_Color newColor) : Graphable(newParent, newColor)
