+++ Frontend/RUGraph/Graphable.cpp
@@ -15,8 +15,8 @@
 // DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include "Graphable.h"
-#include "../../../include/Backend/Database/GList.h"
-#include "../../GFXUtilities/point2.h"
+#include "../../include/Backend/Database/GList.h"
+#include "../GFXUtilities/point2.h"
 #include "RUGraph.h"
 
 Graphable::Graphable(RUGraph* newParent, SDL_Color newColor)
