+++ Frontend/RUGraph/RUDrawComponent.cpp
@@ -15,9 +15,8 @@
 // DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include "RUDrawComponent.h"
-#include "../../../include/Backend/Database/gtype.h"
-#include "../../GFXUtilities/point2.h"
-#include "../../GItems/GPanel.h"
+#include "../../include/Backend/Database/gtype.h"
+#include "../GFXUtilities/point2.h"
 #include "Circle.h"
 #include "GraphLine.h"
 #include "GraphScatter.h"
