+++ Frontend/RUGraph/RUGraph.cpp
@@ -15,11 +15,11 @@
 // DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include "RUGraph.h"
-#include "../../../include/Backend/Database/GList.h"
-#include "../../../include/Backend/Database/gtable.h"
-#include "../../../include/Backend/Database/gtype.h"
-#include "../../GFXUtilities/point2.h"
-#include "../Text/RULabel.h"
+#include "../../include/Backend/Database/GList.h"
+#include "../../include/Backend/Database/gtable.h"
+#include "../../include/Backend/Database/gtype.h"
+#include "../GFXUtilities/point2.h"
+#include "../GUI/Text/RULabel.h"
 #include "GraphLine.h"
 #include "GraphScatter.h"
 #include "Graphable.h"
